+++ app/src/main/java/edu/ucne/jose_estrella_ap2_p2/domain/model/Gasto.kt
@@ -6,6 +6,6 @@ data class Gasto(
     val suplidor : String,
     val ncf: String,
     val itbis: Double,
-    val monto: Int
+    val monto: Double
 
 )
